Split components, Use Vault component with preview prop, fix setpath and create getpath, fix committee function error
