+++ src/components/Vault/Members.scss
@@ -9,6 +9,9 @@
   .twitter-avatar {
     border-radius: 50%;
     width: 40px;
+    height: 40px;
+    object-fit: cover;
+    object-position: center;
   }
   .member-username {
     margin-top: 5px;
