+++ src/components/Vault/Members.tsx
@@ -5,6 +5,7 @@ import "./Members.scss";
 
 interface IProps {
   members: Array<ICommitteeMember>
+  preview?: boolean
 }
 
 export default function Members(props: IProps) {
@@ -20,7 +21,7 @@ export default function Members(props: IProps) {
   const members = props.members?.map((member: ICommitteeMember, index: number) => {
     return (
       <a className="member-link-wrapper" key={index} href={member?.["twitter-link"]} target="_blank" rel="noreferrer">
-        {member?.["image-ipfs-link"] ? <img src={`${IPFS_PREFIX}${verifyIPFSLink(member?.["image-ipfs-link"])}`} alt="twitter avatar" className="twitter-avatar" /> : <TwitterImageIcon />}
+        {member?.["image-ipfs-link"] ? <img src={props.preview ? member?.["image-ipfs-link"] : `${IPFS_PREFIX}${verifyIPFSLink(member?.["image-ipfs-link"])}`} alt="twitter avatar" className="twitter-avatar" /> : <TwitterImageIcon />}
         <span className="member-username">{member.name}</span>
       </a>
     )
