+++ src/components/Vault/Vault.tsx
@@ -1,5 +1,6 @@
 import { useState, useEffect } from "react";
 import "../../styles/Vault/Vault.scss";
+import { t } from "i18next";
 import { IVault } from "../../types/types";
 import { useSelector } from "react-redux";
 import millify from "millify";
@@ -12,15 +13,16 @@ import VaultAction from "./VaultAction";
 
 interface IProps {
   data: IVault,
-  setShowModal: (show: boolean) => any,
-  setModalData: (data: any) => any
+  setShowModal?: (show: boolean) => any,
+  setModalData?: (data: any) => any,
+  preview?: boolean,
 }
 
 export default function Vault(props: IProps) {
   const [toggleRow, setToggleRow] = useState(false);
   const { name, isGuest, bounty, id, description } = props.data;
-  const tokenPrice = useSelector((state: RootState) => state.dataReducer.vaults.filter((vault: IVault) => vault.id === id)[0].parentVault.tokenPrice);
-  const apy = useSelector((state: RootState) => state.dataReducer.vaults.filter((vault: IVault) => vault.id === id)[0].parentVault.apy);
+  const tokenPrice = useSelector((state: RootState) => state.dataReducer.vaults.filter((vault: IVault) => vault.id === id)[0]?.parentVault?.tokenPrice);
+  const apy = useSelector((state: RootState) => state.dataReducer.vaults.filter((vault: IVault) => vault.id === id)[0]?.parentVault?.apy);
   const { totalRewardAmount, honeyPotBalance, withdrawRequests, stakingTokenDecimals } = props.data.parentVault;
   const [vaultAPY, setVaultAPY] = useState("-");
   const [honeyPotBalanceValue, setHoneyPotBalanceValue] = useState("");
@@ -51,7 +53,7 @@ export default function Vault(props: IProps) {
         {formatWei(honeyPotBalance, 3, stakingTokenDecimals)}
         {honeyPotBalanceValue && <span className="honeypot-balance-value">&nbsp;{`≈ $${honeyPotBalanceValue}`}</span>}
       </div>
-      {screenSize === ScreenSize.Mobile && <span className="sub-label">Total vault</span>}
+      {screenSize === ScreenSize.Mobile && <span className="sub-label">{t("Vault.total-vault")}</span>}
     </>
   )
 
@@ -64,7 +66,7 @@ export default function Vault(props: IProps) {
             {/* TODO: handle project-metadata and Project-metadata */}
             <img src={description?.["project-metadata"]?.icon ?? description?.["Project-metadata"]?.icon} alt="project logo" />
             <div className="name-source-wrapper">
-              <div className="project-name">{name}</div>
+              <div className="project-name">{props.preview ? description["project-metadata"].name : name}</div>
               {isGuest && <a className="source-name" target="_blank" rel="noopener noreferrer" href={description?.source?.url}>By {description?.source?.name}</a>}
               {screenSize === ScreenSize.Mobile && maxRewards}
             </div>
@@ -84,7 +86,8 @@ export default function Vault(props: IProps) {
                 data={props.data}
                 withdrawRequests={withdrawRequests}
                 setShowModal={props.setShowModal}
-                setModalData={props.setModalData} />
+                setModalData={props.setModalData}
+                preview={props.preview} />
             </td>
           </>
         )}
@@ -95,7 +98,8 @@ export default function Vault(props: IProps) {
           data={props.data}
           withdrawRequests={withdrawRequests}
           setShowModal={props.setShowModal}
-          setModalData={props.setModalData} />}
+          setModalData={props.setModalData}
+          preview={props.preview} />}
     </>
   )
 }
