+++ src/components/Vault/VaultExpanded.scss
@@ -32,8 +32,9 @@
         justify-content: center;
         margin-top: 50px;
         button {
+          background: transparent;
           color: $drak-green;
-          border-color: $drak-green;
+          border: 1px solid $drak-green;
           width: 350px;
         }
       }
@@ -98,6 +99,9 @@
           max-width: 150px;
           cursor: pointer;
         }
+        button {
+          border: none;
+        }
       }
       span {
         color: $white;
@@ -152,7 +156,7 @@
       min-height: unset;
       overflow: auto;
       white-space: nowrap;
-      width: 90vw;
+      width: calc(100vw - 82px);
     }
     .prize-division-wrapper {
       .pie-chart-wrapper {
