--- src/components/VaultEditor/VaultAction.scss
@@ -1,32 +1,0 @@
-@import "../../styles/constants.scss";
-
-.vault-editor {
-  .vault-action-wrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    > span {
-      font-size: $xxsmall;
-    }
-
-    .deposit-withdraw {
-      border: 1px solid;
-      color: $turquoise;
-      background-color: $light-blue;
-      &:active {
-        color: $white;
-        background-color: $dirty-turquoise;
-        opacity: 1;
-      }
-    }
-  }
-
-  @media only screen and (max-width: $mobile-breakpoint) {
-    .vault-action-wrapper {
-      padding: 20px;
-      .deposit-withdraw {
-        width: 100%;
-      }
-    }
-  }
-}
