--- src/components/VaultEditor/VaultExpanded.scss
@@ -1,185 +1,0 @@
-@import "../../styles/constants.scss";
-
-.vault-editor {
-  .vault-expanded {
-    display: flex;
-    flex-direction: column;
-    text-align: left;
-    .sub-title {
-      font-weight: bold;
-      background-color: $blue;
-      padding: 10px 30px;
-      color: $white;
-      height: 50px;
-      line-height: 50px;
-      font-size: $small;
-    }
-    .vault-details-wrapper {
-      .vault-details-content {
-        display: flex;
-        justify-content: space-between;
-
-        > div {
-          padding: 20px 30px;
-          border: 1px solid $purple-blue;
-          width: 50%;
-        }
-        .multi-sig-wrapper {
-          margin-bottom: 20px;
-        }
-        .submit-vulnerability-button-wrapper {
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          margin-top: 50px;
-          button {
-            background: transparent;
-            color: $drak-green;
-            border: 1px solid $drak-green;
-            width: 350px;
-          }
-        }
-      }
-    }
-    .severity-prizes-wrapper {
-      .severity-prizes-content {
-        display: flex;
-        flex-direction: column;
-      }
-    }
-    .vault-expanded-subtitle {
-      font-style: italic;
-      color: $dark-turquoise;
-      padding-bottom: 20px;
-      display: flex;
-    }
-    .twitter-avatars-wrapper {
-      display: flex;
-      flex-wrap: wrap;
-      max-height: 250px;
-      overflow: scroll;
-    }
-    .severity-prize-division-wrapper {
-      display: flex;
-      flex-direction: column;
-      color: $white;
-      margin-bottom: auto;
-      .division {
-        border-left: 10px solid;
-        padding: 10px 0 10px 15px;
-        &.vested-token {
-          border-left-color: $vested-token;
-        }
-        &.token {
-          border-left-color: $token;
-        }
-        &.committee {
-          border-left-color: $committee;
-        }
-        &.vested-hats {
-          border-left-color: $vested-hats;
-        }
-        &.governance {
-          border-left-color: $governance;
-        }
-        &.swap-and-burn {
-          border-left-color: $swap-and-burn;
-        }
-      }
-    }
-
-    .prize-division-wrapper {
-      display: flex;
-      flex-direction: column;
-      .pie-chart-wrapper {
-        display: flex;
-        align-items: center;
-        height: 100%;
-        .pie-chart-container {
-          svg {
-            max-width: 150px;
-            cursor: pointer;
-          }
-          button {
-            border: none;
-          }
-        }
-        span {
-          color: $white;
-        }
-        .label-wrapper {
-          margin-left: 50px;
-          padding-left: 10px;
-          display: flex;
-          flex-direction: column;
-          border-left: 7px solid;
-          .value {
-            font-weight: bold;
-          }
-        }
-      }
-    }
-
-    .vault-prize {
-      font-size: $medium;
-      font-weight: bold;
-      margin-bottom: 20px;
-      .of-vault-text {
-        color: $white;
-        font-size: $xsmall;
-      }
-    }
-  }
-
-  @media only screen and (max-width: $mobile-breakpoint) {
-    .vault-expanded {
-      .vault-details-wrapper {
-        .vault-details-content {
-          flex-direction: column;
-          padding: 20px;
-          > div {
-            width: 100%;
-            padding: unset;
-            border: unset;
-          }
-          .submit-vulnerability-button-wrapper {
-            button {
-              width: 100%;
-            }
-          }
-        }
-      }
-      .vault-expanded-subtitle {
-        margin-top: 20px;
-      }
-      .twitter-avatars-wrapper {
-        flex-wrap: unset;
-        min-height: unset;
-        overflow: auto;
-        white-space: nowrap;
-        width: 100%;
-      }
-      .prize-division-wrapper {
-        .pie-chart-wrapper {
-          flex-direction: column;
-          .pie-chart-container {
-            display: flex;
-            justify-content: space-between;
-            width: 100%;
-            svg {
-              max-width: 200px;
-            }
-            button {
-              background: none;
-            }
-          }
-          .label-wrapper {
-            margin-top: 20px;
-            margin-left: unset;
-            width: 100%;
-          }
-        }
-      }
-    }
-  }
-}
