+++ src/components/VaultEditor/index.scss
@@ -218,6 +218,115 @@
         }
     }
 
+    // preview vault
+    tr {
+        border-bottom: 2px solid $dark-blue;
+        &.guest {
+            td {
+                background-color: $field-blue !important;
+            }
+        }
+    
+        td {
+            padding: 15px;
+            background-color: $strong-blue;
+            &.sub-row {
+                background-color: transparent;
+                padding: 0px;
+            }
+        }
+    }
+    
+    .arrow {
+        cursor: pointer;
+        transform: rotate(0deg);
+        transition: transform 0.1s linear;
+        margin-left: 20px;
+        &.open {
+            transform: rotate(90deg);
+            transition: transform 0.1s linear;
+        }
+        :hover {
+            opacity: 0.8;
+        }
+    }
+    
+    .project-name-wrapper {
+        display: flex;
+        align-items: center;
+        img {
+            width: 40px;
+            height: 40px;
+            object-fit: cover;
+            object-position: center;
+            margin-right: 15px;
+            border-radius: 50%;
+        }
+        .name-source-wrapper {
+            display: flex;
+            flex-direction: column;
+            text-align: left;
+            .project-name {
+                font-weight: bold;
+            }
+            .source-name {
+                margin-top: 10px;
+                color: $dirty-turquoise;
+            }
+        }
+    }
+    
+    .max-rewards-wrapper {
+        display: flex;
+        text-align: left;
+    }
+    .sub-label {
+        font-size: $tiny;
+        color: $gray;
+    }
+    
+    .honeypot-balance-value {
+        color: $white;
+        font-weight: bold;
+    }
+    
+    .contract-wrapper {
+        display: grid;
+        grid-template-columns: 1fr 1fr;
+        .contract-name {
+            color: $dark-turquoise;
+            overflow: hidden;
+            text-overflow: ellipsis;
+            white-space: nowrap;
+            margin-right: 10px;
+        }
+        .contract-value {
+            max-width: 150px;
+            overflow: hidden;
+            text-overflow: ellipsis;
+            white-space: nowrap;
+        }
+    }
+    
+    .contracts-covered-modal-wrapper {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        padding: 50px 0;
+        border-top: 1px solid $dark-turquoise;
+        > a {
+            width: 60%;
+            display: flex;
+            justify-content: space-between;
+        }
+    }
+
+    .preview-vault {
+        margin-bottom: 30px;
+        border: 1px solid $turquoise;
+        box-sizing: border-box;
+    }
+
     @media only screen and (max-width: $mobile-breakpoint) {
         margin-left: 20px !important;
         margin-right: 20px !important;
@@ -364,6 +473,24 @@
         }
         .desktop-only {
             display: none;
+        }
+
+        // preview vault
+        .arrow {
+            margin-left: unset;
+        }
+        .rewards-cell {
+            width: 100%;
+        }
+        .contracts-covered-modal-wrapper {
+            padding: 40px 20px;
+            > a {
+                width: 100%;
+            }
+        }
+        .max-rewards-wrapper {
+            flex-direction: row;
+            margin: 5px 0;
         }
     }
 }
